test: fold DashboardMergedGating KAT into allowlisted honesty target (#769 Not-Run trap)
The 7 DashboardMergedGating cases lived in a standalone add_executable
(test_dashboard_merged_gating) absent from the build.yml --target
allowlist, so CI never compiled it: ctest registered the cases from a
source scan and reported them (Not Run) -> exit 8, and the allowlist
drift-guard failed closed on exactly that drift. This is the #769 trap.

Fold the cases (pure static-HTML characterization, no core/ltc link
needed) into test_web_honesty_regression -- an already-allowlisted target
covering the same charter (the dashboard must never imply revenue that
does not exist, per #896). Add the DASHBOARD_HTML_PATH compile-def to that
target; delete the standalone target and source file. No new CMake target,
no build.yml edit.

Verified now-built + non-vacuous: all 7 PASS on fixed HTML (folded suite
49/49); all 7 FAIL on the pre-fix HTML (git merge-base:web-static/
dashboard.html), including NoHardcodedMergedSymbolOnBlockValue against the
operator-reported +<merged_block_value> DOGE card (dashboard.html:1852).
