+++ test/CMakeLists.txt
@@ -1333,13 +1333,8 @@ if (BUILD_TESTING AND GTest_FOUND)
     )
     target_link_libraries(test_web_honesty_regression PRIVATE c2pool_payout c2pool_merged_mining c2pool_hashrate c2pool_storage ltc_coin pool)  # OBJECT-lib SCC direct-naming (#22/#39)
     gtest_add_tests(test_web_honesty_regression "" AUTO)
-
-    add_executable(test_dashboard_merged_gating test_dashboard_merged_gating.cpp)
-    target_compile_definitions(test_dashboard_merged_gating PRIVATE
+    target_compile_definitions(test_web_honesty_regression PRIVATE  # folded #912 static-HTML KAT
         DASHBOARD_HTML_PATH="${CMAKE_SOURCE_DIR}/web-static/dashboard.html")
-    target_link_libraries(test_dashboard_merged_gating PRIVATE
-        GTest::gtest_main GTest::gtest)
-    gtest_add_tests(test_dashboard_merged_gating "" AUTO)
 
 endif()
